Innitial batch of repository code files.  These files have been slightly updated  and debugged vs the dissertation files, and are incomplete.
The repo will load files containing the meandering river centerline evolution simulation and analyze and depict the planform dynamics, and calculate the storage time distributions.

Not yet included are the code files which will calculate a functional fit for the storage time distributions.  These files are to follow, but I need access to the Matlab Curve Fitting Toolbox to debug the code before posting.